+++ packages/web-overlays/src/tooltip/TooltipProps.ts
@@ -21,6 +21,10 @@ export type TooltipProps = {
    * @default 4
    * */
   zIndex?: PositionStyles['zIndex'];
+  /**
+   * A unique ID used to ensure tooltips are accessible
+   */
+  tooltipId?: string;
 } & TooltipBaseProps;
 
 export type PopperTooltipProps = {