+++ packages/web-overlays/src/tooltip/useTooltipState.ts
@@ -1,11 +1,12 @@
-import { useCallback, useId, useMemo } from 'react';
+import { useCallback, useMemo } from 'react';
 import { useToggler } from '@cbhq/cds-common';
+import { usePrefixedId } from '@cbhq/cds-common/hooks/usePrefixedId';
 
-export const useTooltipState = () => {
+export const useTooltipState = (id?: string) => {
   const [isHovered, { toggleOn: handleOnMouseEnter, toggleOff: toggleOffIsHovered }] =
     useToggler(false);
   const [isFocused, { toggleOn: handleOnFocus, toggleOff: toggleOffIsFocused }] = useToggler(false);
-  const tooltipId = useId();
+  const tooltipId = usePrefixedId(id);
 
   const handleOnBlur = useCallback(() => {
     toggleOffIsFocused();